+++ server/fleet/datastore.go
@@ -2412,8 +2412,6 @@ type Datastore interface {
 
 	// GetCurrentTime gets the current time from the database
 	GetCurrentTime(ctx context.Context) (time.Time, error)
-
-	InsertInHouseApp(ctx context.Context, payload *InHouseAppPayload) (installerID uint, titleID uint, err error)
 }
 
 type AndroidDatastore interface {
